--- backend/lambdas/offers/createOffersTable/package.json
@@ -1,19 +1,0 @@
-{
-  "name": "offers-createtable",
-  "version": "1.0.0",
-  "description": "One time execution lambda to create (if not exists) the offers table",
-  "main": "index.js",
-  "scripts": {
-    "compile": "npx tsc"
-  },
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "aws-sdk": "^2.1691.0",
-    "pg": "^8.13.0"
-  },
-  "devDependencies": {
-    "@types/pg": "^8.11.10",
-    "typescript": "^5.6.3"
-  }
-}
