--- backend/lambdas/offers/getHighestOffer/package.json
@@ -1,20 +1,0 @@
-{
-  "name": "gethighestoffer",
-  "version": "1.0.0",
-  "main": "index.js",
-  "scripts": {
-    "compile": "npx tsc"
-  },
-  "author": "",
-  "license": "ISC",
-  "description": "",
-  "dependencies": {
-    "aws-sdk": "^2.1691.0",
-    "pg": "^8.13.0"
-  },
-  "devDependencies": {
-    "@types/aws-lambda": "^8.10.145",
-    "@types/pg": "^8.11.10",
-    "typescript": "^5.6.2"
-  }
-}
