--- backend/lambdas/offers/getHighestOffer/tsconfig.json
@@ -1,13 +1,0 @@
-{
-    "compilerOptions": {
-      "module": "commonjs",
-      "esModuleInterop": true,
-      "target": "es6",
-      "moduleResolution": "node",
-      "sourceMap": true,
-      "outDir": "dist"
-    },
-    "lib": ["es2015"],
-    "include": ["./**/*.ts"],
-    "exclude": ["node_modules"]
-}
