+++ backend/lambdas/offers/tsconfig.json
